Add upload script
It's a hassle to manually upload new builds
of the plugin via Google's web UI.
This will let you upload a newer version of
the chrome extension programmatically:

node build.js
node upload.js

For this to work, you need the client secret,
and a refresh token. Documentation for that will
come later.